# increase decode_varint performance