@ajguerrer 2021-09-06
Initial commit was wrong and could invoke UB with `get_unchecked` (forgot I was working with `&[u8]` and not `Bytes`, which act like a cursor). Updated the PR description with new performance numbers. Unfortunately it made the perf improvements drop a bit, but hopefully still worth merging.

I thought about writing a unit test to capture my mistake, but the problem with invoking UB is that my test was passing for all the wrong reasons, so it ended up not being very useful.