@jaskij 2021-11-02
You can not get rid of the slow decode loop, unfortunately.

`bytes::buf::Buf` documentation states that:
> The underlying storage may or may not be in contiguous memory.

Also, for `chunk()`:
> Returns a slice starting at the current position and of length between 0 and `Buf::remaining()`.

So you cannot rely on all bytes of the varint being in the slice returnd by `chunk()`. They could be spread among multiple chunks, depending on the actual implementation of the trait. This would be a subtle bug, which could or could not come up in the future, depending on what actual `Buf` implementations are used.

(For this same reason the default implementation of `Buf::chunks_vectored` seems buggy to me).