@ajguerrer 2021-11-04
@jaskij Thanks for pointing that out. The naming behind `chunk()` makes more sense to me now too. I am going to go ahead and close this PR since it would not survive the tests added by #559.